reduce compiler warnings

+++ jaro_winkler/details/intrinsics.hpp
@@ -16,7 +16,7 @@ template <typename T>
 T bit_mask_lsb(int n)
 {
     T mask = -1;
-    if (n < sizeof(T) * 8) {
+    if (n < (int)(sizeof(T) * 8)) {
         mask += (T)1 << n;
     }
     return mask;
@@ -28,7 +28,7 @@ bool bittest(T a, int bit)
     return (a >> bit) & 1;
 }
 
-static inline size_t popcount64(uint64_t x)
+static inline int64_t popcount64(uint64_t x)
 {
     const uint64_t m1 = 0x5555555555555555;
     const uint64_t m2 = 0x3333333333333333;
@@ -38,7 +38,7 @@ static inline size_t popcount64(uint64_t x)
     x -= (x >> 1) & m1;
     x = (x & m2) + ((x >> 2) & m2);
     x = (x + (x >> 4)) & m4;
-    return static_cast<size_t>((x * h01) >> 56);
+    return static_cast<int64_t>((x * h01) >> 56);
 }
 
 /**
